+++ s3sup/catalogue.py
@@ -1,6 +1,7 @@
 import csv
 import enum
 import collections
+import copy
 import click
 import inflect
 
@@ -10,7 +11,8 @@ class ChangeReason(enum.Enum):
     CONTENT_CHANGED = 2
     ATTRIBUTES_CHANGED = 3
     DELETED = 4
-    NO_CHANGE = 5
+    DELETED_PROTECTED = 5
+    NO_CHANGE = 6
 
 
 CR_STYLE = collections.namedtuple(
@@ -23,14 +25,17 @@ CR_STYLES = {
     ChangeReason['ATTRIBUTES_CHANGED']: CR_STYLE(
         'cyan', '~', 'attrs', 'attributes changed'),
     ChangeReason['DELETED']: CR_STYLE('red', '-', 'delete', 'deleted'),
+    ChangeReason['DELETED_PROTECTED']: CR_STYLE(
+        'green', '^', 'deleteprotected', 'deleted but protected'),
     ChangeReason['NO_CHANGE']: CR_STYLE('green', '^', 'unchanged', 'unchanged')
 }
 
 
 class Catalogue:
 
-    def __init__(self):
+    def __init__(self, preserve_deleted_files=False):
         self._c = {}
+        self._preserve_deleted_files = preserve_deleted_files
 
     def add_file(self, path: str, content_hash: str, attributes_hash: str):
         self._c[path] = (str(content_hash), str(attributes_hash))
@@ -56,6 +61,8 @@ class Catalogue:
     def diff_dict(self, remote_catalogue):
         lcl = self.to_dict()
         rmt = remote_catalogue.to_dict()
+        # TODO: Not a great use of memory, see if there's a better way.
+        new_rmt = copy.deepcopy(self)
 
         changes = {
             'num_changes': 0,
@@ -64,27 +71,40 @@ class Catalogue:
                 'content_changed': [],
                 'attributes_changed': []
             },
-            'delete': [pth for pth in rmt if pth not in lcl],
+            'delete': [],
+            'delete_protected': [],
             'unchanged': []
         }
-        changes['num_changes'] += len(changes['delete'])
+
+        for path, (content_hash, attributes_hash) in rmt.items():
+            if path not in lcl:
+                if self._preserve_deleted_files:
+                    changes['delete_protected'].append(path)
+                    # Deleted but protected files are the only ones that need
+                    # to be maintained in remote
+                    new_rmt.add_file(path, content_hash, attributes_hash)
+                else:
+                    changes['delete'].append(path)
 
         for path, (content_hash, attributes_hash) in lcl.items():
             try:
                 r_content_hash, r_attributes_hash = rmt[path]
                 if content_hash != r_content_hash:
                     changes['upload']['content_changed'].append(path)
-                    changes['num_changes'] += 1
                 elif attributes_hash != r_attributes_hash:
                     changes['upload']['attributes_changed'].append(path)
-                    changes['num_changes'] += 1
                 else:
                     changes['unchanged'].append(path)
             except KeyError:
                 changes['upload']['new_files'].append(path)
-                changes['num_changes'] += 1
 
-        return changes
+        changes['num_changes'] = (
+            len(changes['delete'])
+            + len(changes['upload']['new_files'])
+            + len(changes['upload']['content_changed'])
+            + len(changes['upload']['attributes_changed'])
+        )
+        return changes, new_rmt
 
 
 def print_diff_summary(dd, verbose=False):
@@ -128,22 +148,27 @@ def print_diff_summary(dd, verbose=False):
     _p(dd['upload']['content_changed'], 'CONTENT_CHANGED')
     _p(dd['upload']['attributes_changed'], 'ATTRIBUTES_CHANGED')
     _p(dd['delete'], 'DELETED')
+    _p(dd['delete_protected'], 'DELETED_PROTECTED')
     _p(dd['unchanged'], 'NO_CHANGE')
 
 
 def change_list(diff):
     """
-    Return only paths that need changing, in a list with format:
-    TODO
+    Paths that need changes made on S3, along with the reason.
+
+    Changes should be made in the order returned, in an attempt to prevent the
+    temporary period where there may be incorrect links between HTML files and
+    static assets that haven't been uploaded yet.
     """
-    dl = [(ChangeReason.NEW_FILE, p)
-          for p in diff['upload']['new_files']]
+    # Attribute changes first, less risk of content changes.
+    dl = [(ChangeReason.ATTRIBUTES_CHANGED, p)
+          for p in diff['upload']['attributes_changed']]
+    # TODO: Upload in order: others, CSS, JS, HTML
+    dl += [(ChangeReason.NEW_FILE, p)
+           for p in diff['upload']['new_files']]
+    # TODO: Upload in same order as above
     dl += [(ChangeReason.CONTENT_CHANGED, p)
            for p in diff['upload']['content_changed']]
-    dl += [(ChangeReason.ATTRIBUTES_CHANGED, p)
-           for p in diff['upload']['attributes_changed']]
-    dl += [(ChangeReason.ATTRIBUTES_CHANGED, p)
-           for p in diff['upload']['attributes_changed']]
     dl += [(ChangeReason.DELETED, p)
            for p in diff['delete']]
     return dl