+++ tests/test_catalogue.py
@@ -113,7 +113,7 @@ class TestCatalogueDiff(unittest.TestCase):
             .add_file('robots.txt', 'asdfhl', 'lkjfds')
             .add_file('tempfile.txt', 'fj8fj8', 'flwlfwl')
         )
-        diff_dict = local_cat.diff_dict(remote_cat)
+        diff_dict, new_remote_catalogue = local_cat.diff_dict(remote_cat)
         expected = {
             'num_changes': 5,
             'upload': {
@@ -122,9 +122,47 @@ class TestCatalogueDiff(unittest.TestCase):
                 'attributes_changed': ['assets/blam/160-180.jpg']
             },
             'delete': ['tempfile.txt'],
+            'delete_protected': [],
             'unchanged': ['consistent.html.html', '♬ /music.fav.mp3']
         }
         self.assertEqual(expected, diff_dict)
+
+    def test_diff_dict_with_protected_deletion(self):
+        local_cat = (
+            Catalogue(preserve_deleted_files=True)
+            .add_file('index.html', '9J9J9J', 'P2P2P2')
+            .add_file('assets/blam/160-180.jpg', 'A1A1A1', 'B3B3B3')
+            .add_file('robots.txt', '4b4b4b', '929292')
+            .add_file('consistent.html.html', '123', '123')
+            .add_file('news_update.html', '4b4b4b', '929292')
+        )
+        remote_cat = (
+            Catalogue()
+            .add_file('assets/blam/160-180.jpg', 'A1A1A1', '9S9S95')
+            .add_file('consistent.html.html', '123', '123')
+            .add_file('index.html', '282828', 'P2P2P2')
+            .add_file('♬ /music.fav.mp3', 200010, '7A9 ')
+            .add_file('robots.txt', 'asdfhl', 'lkjfds')
+            .add_file('tempfile.txt', 'fj8fj8', 'flwlfwl')
+        )
+        diff_dict, new_remote_catalogue = local_cat.diff_dict(remote_cat)
+        expected_diff = {
+            'num_changes': 4,
+            'upload': {
+                'new_files': ['news_update.html'],
+                'content_changed': ['index.html', 'robots.txt'],
+                'attributes_changed': ['assets/blam/160-180.jpg']
+            },
+            'delete': [],
+            'delete_protected': ['tempfile.txt', '♬ /music.fav.mp3'],
+            'unchanged': ['consistent.html.html']
+        }
+        self.assertEqual(expected_diff, diff_dict)
+        rmt_cat_d = new_remote_catalogue.to_dict()
+        self.assertTrue('tempfile.txt' in rmt_cat_d)
+        self.assertEqual(('fj8fj8', 'flwlfwl'), rmt_cat_d['tempfile.txt'])
+        self.assertTrue('♬ /music.fav.mp3' in rmt_cat_d)
+        self.assertEqual(('200010', '7A9 '), rmt_cat_d['♬ /music.fav.mp3'])
 
 
 if __name__ == '__main__':